Refactor workflow utils to improve API and organization
Simplify buildWorkflowFromNode to accept arrays instead of maps, extract
test data to dedicated file, and split functionality into separate files
with clear responsibilities.
