+++ packages/workflow-utils/src/index.ts
@@ -1,1 +1,2 @@
-export * from "./build-workflow";
+export * from "./build-workflow-map";
+export * from "./build-workflow-from-node";
